finalizado el proyecto de hospital

+++ HospitalProyectBack/Models/Client.cs
@@ -4,7 +4,7 @@
     {   
         public int Id { get; set; }
         public string Name { get; set; }
-        public long Phone { get; set; }
+        public int Phone { get; set; }
         public int Dni { get; set; }
         public string Observation { get; set; }
         public string Email;
